backend/requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.42.1
idna==3.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
